Fix an issue where the score would cap at 65535

+++ FlockingHell/FlockingHell/Source/Public/Player.h
@@ -43,8 +43,8 @@ struct Player
 	unsigned short BossKilled = 0;
 	unsigned short EnemiesKilled = 0;
 	unsigned short BulletDamage = 0;
-	unsigned short Score = 0;
-	unsigned short GrazingScore = 0;
+	unsigned int Score = 0;
+	unsigned int GrazingScore = 0;
 
 	bool bInvincible{};
 	bool bIsHit{};
